fix: do not require to unset PR environment variables
