+++ cli.js
@@ -4,6 +4,7 @@
 const proxyquire = require('proxyquire');
 const { name, version } = require('./package.json');
 const semantishPrerelease = require('./src');
+const hidePr = require('./src/hidePr');
 
 console.log(`${name}@${version}`);
 
@@ -11,6 +12,7 @@ proxyquire('semantic-release/cli', {
   '.': (opts) => {
     return semantishPrerelease(opts);
   },
+  'env-ci': hidePr(process.argv.includes('--release-pr')),
 })().catch((err) => {
   console.error(err);
   process.exit(err.code || 1);
